Изменение номер 1

+++ src/ru/mpts/units/TaskPlayers.java
@@ -164,12 +164,12 @@ public abstract class TaskPlayers {
                 Engine.graphics2D.setColor(new Color(0xAC5800));
             } else {
                 Engine.graphics2D.setColor(new Color(0x00A8BB));
-                Engine.graphics2D.fillRect((int) (taskAction.get(i).getLocation().getX() * Map.getScale() + Map.getIndentX() + 2), (int) (taskAction.get(i).getLocation().getY() * Map.getScale() + Map.getIndentY() + 2),
-                        (int) (Map.getScale() - 5), (int) (Map.getScale() - 5));
+                Engine.graphics2D.fillRect(taskAction.get(i).getLocation().getX() * Map.getScale() + Map.getIndentX() + 2, taskAction.get(i).getLocation().getY() * Map.getScale() + Map.getIndentY() + 2,
+                        Map.getScale() - 5, Map.getScale() - 5);
                 Engine.graphics2D.setColor(new Color(0xAF0063));
             }
-            Engine.graphics2D.drawRect((int) (taskAction.get(i).getLocation().getX() * Map.getScale() + Map.getIndentX() + 2), (int) (taskAction.get(i).getLocation().getY() * Map.getScale() + Map.getIndentY() + 2),
-                    (int) (Map.getScale() - 5), (int) (Map.getScale() - 5));
+            Engine.graphics2D.drawRect(taskAction.get(i).getLocation().getX() * Map.getScale() + Map.getIndentX() + 2, taskAction.get(i).getLocation().getY() * Map.getScale() + Map.getIndentY() + 2,
+                    Map.getScale() - 5, Map.getScale() - 5);
 
         }
     }
